Plan: Não permitir deletar planos que possuem detalhes

+++ app/Http/Controllers/Admin/PlanController.php
@@ -77,10 +77,19 @@ class PlanController extends Controller
 
     public function destroy($id)
     {
-        $plan = $this->repository->where('id', $id)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('id', $id)
+            ->first();
 
         if (!$plan) {
             return redirect()->back();
+        }
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detalhes vinculados a esse plano, portanto não pode deletar.');
         }
 
         $plan->delete();
